Changes
## Added new Half-Life/binaries folder.
This folder contains all the game files that need to be added so the end user can just install the port, copy the files from the steam install, then on first run it will put the right files in the right place replacing the necessary files.

## Added build.py
This script pieces the whole thing together, place the results of building the xash3d/hlsdk-portable in the build/ folder and this should produce a file suitable for Portmaster.

## Various Fixes

- Added liblist.gam: to stop xash3d from trying to import hl_arm64_arm64.so
- Removed video.cfg: this should autogenerate
- Added various files to support auto copying of files from the binaries folder.